Add more test cases, and refactory.
