+++ src/LineBot/ILineBot.cs
@@ -24,6 +24,13 @@ namespace Line
     public interface ILineBot
     {
         /// <summary>
+        /// Creates a rich menu.
+        /// </summary>
+        /// <param name="richMenu">The rich menu represented as a rich menu object.</param>
+        /// <returns>.</returns>
+        Task<string> CreateRichMenu(RichMenu richMenu);
+
+        /// <summary>
         /// Returns the content of the specified message.
         /// </summary>
         /// <param name="message">The message.</param>
@@ -217,12 +224,5 @@ namespace Line
         /// <param name="messages">The messages to send.</param>
         /// <returns>.</returns>
         Task<ILineBot> Reply(string replyToken, params ISendMessage[] messages);
-
-        /// <summary>
-        /// Creates a rich menu.
-        /// </summary>
-        /// <param name="richMenu">The rich menu represented as a rich menu object.</param>
-        /// <returns>.</returns>
-        Task<string> CreateRichMenu(RichMenu richMenu);
     }
 }
