+++ src/LineBot/RichMenu/RichMenuResponse.cs
@@ -22,7 +22,7 @@ namespace Line
     public sealed class RichMenuResponse : RichMenu
     {
         /// <summary>
-        /// Gets or sets Rich menu ID.
+        /// Gets or sets the rich menu ID.
         /// </summary>
         [JsonProperty("richMenuId")]
         public string RichMenuId { get; set; }
